reduces ping
added the functions that code asked for. i misunderstood his concerns and was totally wrong about it requiring a total restructure. i'll get around to the restructure one of these days... 

the if statements are one by one to avoid unnecessary string comparisons on the assumption that more people are DG/SG than DCG/SCG